requirements.txt
azure-functions
azure-core
python-dotenv
openai
